Backend/requirements.txt
Flask==0.12.2
gunicorn==19.7.1
google-cloud-vision==0.28.0
google-cloud-storage==1.6.0
google-cloud-datastore==1.4.0
Pillow==4.3.0
Pyrebase==3.0.27
firebase-admin==2.5.0
